spec: make i18n selenium specs work with webpack
refs CNVS-25916

also makes some env var conditionals more
robust, as docker-compose env vars that are booleans
get translated as init-capped strings

TEST PLAN:
 1) i18n_js_spec.rb should pass in webpack selenium
    build

Reviewed-on: https://gerrit.instructure.com/69333
Tested-by: Jenkins
Product-Review: Ethan Vizitei <evizitei@instructure.com>
QA-Review: Ethan Vizitei <evizitei@instructure.com>

+++ lib/tasks/js.rake
@@ -70,7 +70,7 @@ namespace :js do
   desc 'test javascript specs with Karma'
   task :test, :reporter do |task, args|
     reporter = args[:reporter]
-    if ENV['USE_WEBPACK'].present? && ENV['USE_WEBPACK'] != 'false'
+    if ENV['USE_WEBPACK'].present? && ENV['USE_WEBPACK'] != 'false' && ENV['USE_WEBPACK'] != 'False'
       Rake::Task['i18n:generate_js'].invoke
       webpack_test_dir = Rails.root + "spec/javascripts/webpack"
       FileUtils.rm_rf(webpack_test_dir)
@@ -300,7 +300,7 @@ namespace :js do
       else
         puts "--> Building DEVELOPMENT webpack bundles"
         `npm run webpack-development`
-        if ENV['USE_OPTIMIZED_JS'] == 'true'
+        if ENV['USE_OPTIMIZED_JS'] == 'true' || ENV['USE_OPTIMIZED_JS'] == 'True'
           # if this var is set, we'll need to have optimized version of the
           # webpack bundles available too
           puts "--> Building OPTIMIZED webpack bundles"
